+++ modules/azure/linux-vm/kv.tf
@@ -1,9 +1,4 @@
-#tfsec:ignore:AZU020 tfsec:ignore:AZU021
 resource "tls_private_key" "ssh_admin_key" {
-  for_each = {
-    for name in var.vm_config :
-    vm_config.name => name
-  }
   algorithm = "RSA"
   rsa_bits  = "4096"
 }
@@ -11,11 +6,11 @@ resource "tls_private_key" "ssh_admin_key" {
 #tfsec:ignore:AZU023
 resource "azurerm_key_vault_secret" "ssh_admin_key_secret" {
   for_each = {
-    for name in var.vm_config :
-    vm_config.name => name
+    for vm in var.vm_config :
+    vm.vm_name => vm
   }
-  name         = var.vm_config[each.key].ssh_key_name
-  value        = trimspace(tls_private_key.ssh_admin_key[each.key].private_key_openssh)
-  key_vault_id = var.vm_config[each.key].key_vault_id
+  name         = "ssh-key-${var.environment}-${var.location_short}-${each.value.vm_name}"
+  value        = trimspace(tls_private_key.ssh_admin_key.private_key_openssh)
+  key_vault_id = each.value.key_vault_id
   content_type = "x509 string"
 }