+++ validation/azure/linux-vm/main.tf
@@ -23,6 +23,8 @@ module "linux_adextension" {
       rg_name              = "rg-dev-we-linux"
       subnet_id            = "value"
       key_vault_id         = "test"
+      ssh_key_name         = "test"
+      ip_config_name       = "test"
     },
     {
       name                 = "vm-02"
@@ -40,23 +42,8 @@ module "linux_adextension" {
       rg_name              = "rg-dev-we-linux"
       subnet_id            = "value"
       key_vault_id         = "test"
-    },
-    {
-      name                 = "vm-03"
-      size                 = "test"
-      publisher            = "test"
-      offer                = "test"
-      sku                  = "test"
-      version              = "test"
-      caching              = "test"
-      storage_account_type = "SSD"
-      disk_size_gb         = 128
-      username             = "localadmin"
-      nic_name             = "vm-nic"
-      location             = "West Europe"
-      rg_name              = "rg-dev-we-linux"
-      subnet_id            = "value"
-      key_vault_id         = "test"
+      ssh_key_name         = "test"
+      ip_config_name       = "test"
     },
   ]