+++ CHANGELOG.md
@@ -8,5 +8,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-- [#2](https://github.com/XenitAB/cloudops-terraform-modules/pull/2) Refactor module to work with Make.
+- [#3](https://github.com/XenitAB/cloudops-terraform-modules/pull/3) Refactor for_each to work better.
+- [#2](https://github.com/XenitAB/cloudops-terraform-modules/pull/2) Refactor module to work with Make and use For_Each instead of count.